# Conflicts:
#	notebooks/preprocess-adapt-aviris.ipynb

+++ spectraclass/ext/umap/umap_.py
@@ -2382,9 +2382,9 @@ class UMAP(BaseEstimator):
                 raise ValueError("Non-zero distances from samples to themselves!")
             if self.knn_dists is None:
                 self._knn_indices = np.zeros(
-                    (X.shape[0], self.n_neighbors), dtype=np.int
+                    (X.shape[0], self.n_neighbors), dtype=np.int32
                 )
-                self._knn_dists = np.zeros(self._knn_indices.shape, dtype=np.float)
+                self._knn_dists = np.zeros(self._knn_indices.shape, dtype=np.float32)
                 for row_id in range(X.shape[0]):
                     # Find KNNs row-by-row
                     row_data = X[row_id].data
